[!!!][CLEANUP] Class names and class files
This change is quite breaking if other extensions
used the SchemaService. However the classes are not
really considered as API, but only the commands.

Besides that, we're heading towards a new major version,
where breaking changes are acceptable
